Added a highScore variable, and some methods for answering questions
Added a highScore variable, which can be saved using SharedPreferences.
Added a curQuestion variable to increment each time player answers a
question.
Added a couple methods to determine if a level is over, and if the game
is over.
Created a loadNextQuestion method which removes a Question from the
Quiz, and inserts the text in the questionTextView and answerButtons.