Lots more functionality added.
- Changed activity_level_start.xml layout, and LevelStart.java, so that
the background image is added programmatically depending on which level
the player is going to start.
- Added onClick handler for the answer buttons, which handles
determining if the player was right/wrong, and whether to move to the
next level, move to the game over screen, or continue to the next
question.
- Added code that updates the icons at the bottom of the screen if the
player gets the question correct or incorrect.
- Added temporary images to res/drawable-mdpi folder to display between
levels. Will just need to overwrite these with the finalized images
(level1.png, level2.png, level3.png).